requirements.txt
fastapi
uvicorn
pydantic
yt-dlp
openai-whisper
pydub
ffmpeg-python